+++ src/main/java/com/movie4u/sep/models/Movie.java
@@ -12,7 +12,7 @@ public class Movie {
 
     private int id;
     private String title;
-    private int year;
+    private Long year;
 
     private Rating rating;
